+++ src/components/index.ts
@@ -2,11 +2,12 @@ import Planner from './planner';
 import AppHeader from './app_header';
 import EditHeader from './edit_header';
 import GoalTracker from './GoalTracker';
-import AddPlanDeatails from './add_plan_details';
+import AddPieceContainer from "./AddPiece/AddPieceContainer"
+import AddPlanDeatails from './AddPiece/add_plan_details';
 import ViewPlanDetails from './view_plan_details';
 import DropdownCalendar from './dropdown_calendar';
 import DropdownSelector from './dropdown_selector';
 import ProfileLogoSection from './profile_logo_section';
 
 
-export { Planner, AppHeader, EditHeader, GoalTracker, AddPlanDeatails, ViewPlanDetails, DropdownCalendar, DropdownSelector, ProfileLogoSection };
+export { Planner, AppHeader, EditHeader, GoalTracker, AddPlanDeatails, ViewPlanDetails, DropdownCalendar, DropdownSelector, ProfileLogoSection,AddPieceContainer };
